all clang-tidys

+++ asm_x64.c
@@ -396,12 +396,12 @@ void emit_jmp32_patch_here(emit_state *s, int64_t patch) {
 // Emit R15 twice, so we have an odd number - and therefore a balanced stack
 uint8_t callee_save[] = {RBX, RBP, R12, R13, R14, R15, R15};
 void restore_callee_regs(emit_state *s) {
-  for (uint8_t i = 0; i < ARRAY_LEN(callee_save); i++) {
+  for (size_t i = 0; i < ARRAY_LEN(callee_save); i++) {
     emit_pop(s, callee_save[i]);
   }
 }
 void save_callee_regs(emit_state *s) {
-  for (uint8_t i = ARRAY_LEN(callee_save); i > 0; i--) {
+  for (size_t i = ARRAY_LEN(callee_save); i > 0; i--) {
     emit_push(s, callee_save[i - 1]);
   }
 }
